code and database update. * UCS now uses the new stream. * update net/eqstream.h and net/eqstream.cpp (imported from eqemu) * verify mailkey now works in UCS. * fixes consent messages for consenter. * add UCSServerStatus (imported from eqemu) * usc messages for connecting / disconnecting. * add zone connecting messages * change to cle enum. (import from eqemu) * add rule based classic ZEM experience * add rule based classic items stats. * fix gm command crash with empty argument * made merchant NPCs Immune to charm. * add CheckForPossibleConfigurationIssues (import from eqemu)

+++ queryserv/worldserver.cpp
@@ -43,7 +43,9 @@ WorldServer::WorldServer()
 {
 }
 
-WorldServer::~WorldServer() {}
+WorldServer::~WorldServer() 
+{
+}
 
 void WorldServer::Connect()
 {
@@ -53,8 +55,7 @@ void WorldServer::Connect()
 		false,
 		"QueryServ",
 		Config->SharedKey
-	); 
-	
+		);
 	m_connection->OnMessage(std::bind(&WorldServer::HandleMessage, this, std::placeholders::_1, std::placeholders::_2));
 }
 
@@ -184,7 +185,7 @@ void WorldServer::HandleMessage(uint16 opcode, const EQ::Net::Packet& p)
 			ServerPacket pack;
 			pack.pBuffer = (uchar*)p.Data();
 			pack.opcode = opcode;
-			pack.size = p.Length();
+			pack.size = (uint32)p.Length();
 
 			database.GeneralQueryReceive(&pack);
 			pack.pBuffer = nullptr;
@@ -192,4 +193,3 @@ void WorldServer::HandleMessage(uint16 opcode, const EQ::Net::Packet& p)
 		}
 	}
 }
-
